feat: parallel multi-doc parse (item 4) + CST u32 compression (item 5)
## Item 4 — Parallel multi-document parsing

Adds `noyalib::load_all_as_parallel<T>` and
`noyalib::load_all_parallel` behind a `parallel` feature flag.
Pre-scans `---` document boundaries on a single thread, then
deserialises each document in parallel via Rayon. Targets
multi-document streams (telemetry logs, audit exports,
Kubernetes-resource snapshots) where single-thread parsing is
CPU-bound.

Public surface:
  - `load_all_as_parallel<T>(input) -> Result<Vec<T>>`
  - `load_all_parallel(input) -> Result<Vec<Value>>`
  - `parallel::split_documents(input) -> Vec<&str>` (standalone
    boundary pre-scanner for callers that want their own
    concurrency primitives).

10 unit tests covering boundary detection edge cases (no
separators, empty input, implicit first document, mid-line
`---`, dashes followed by non-whitespace) and end-to-end
correctness — `load_all_as_parallel` results match
`load_all_as` bit-for-bit on a 50-document stress test.

## Item 5 — CST short-pointer compression

`GreenChild::Token { len }` and `GreenNode.text_len` are now
`u32` (were `usize`). YAML documents are bounded at 4 GiB by the
parser's `max_document_length` cap, so a `u32` is sufficient.

The narrower field drops `GreenChild::Token` from 24 bytes to 8
bytes on a 64-bit target — meaningfully better L1/L2 cache
locality on tree traversals. For a CST with 10k tokens this
saves 160 KB of cache footprint.

Public surface:
  - `GreenChild::Token { kind, len: u32 }` is now `u32` instead
    of `usize`. Pattern matches that used `len` in slice
    arithmetic need `*len as usize`. The doctest example
    shows the canonical pattern.
  - `text_len()` accessors on `GreenChild` and `GreenNode`
    preserve their `usize` return type — the narrower storage
    is widened at the API boundary so existing callers continue
    to compile without changes.

`GreenNode::new` now `u32::try_from`s the summed text_len with
an `expect` panic message; the parser's `max_document_length`
cap (default 64 MiB, max 4 GiB) guarantees the conversion
never fires.

## Verification

- 406/406 YAML 1.2 official test suite still passes.
- 363 main-crate tests, 22 workspace member tests — all green.
- `cargo clippy --workspace --all-features --all-targets`: clean.
- `cargo +1.75.0 check --locked`: MSRV gate clean.

---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co

+++ src/cst/format.rs
@@ -147,7 +147,7 @@ impl<'a> Formatter<'a> {
                     self.format_node(inner, pos)?;
                 }
                 GreenChild::Token { kind, len } => {
-                    self.handle_token(*kind, &self.source[pos..pos + *len]);
+                    self.handle_token(*kind, &self.source[pos..pos + *len as usize]);
                 }
             }
             pos += child.text_len();
@@ -205,7 +205,7 @@ impl<'a> Formatter<'a> {
                     self.format_mapping_entry(inner, pos)?;
                 }
                 GreenChild::Token { kind, len } => {
-                    self.handle_token(*kind, &self.source[pos..pos + *len]);
+                    self.handle_token(*kind, &self.source[pos..pos + *len as usize]);
                 }
                 GreenChild::Node(inner) => {
                     self.format_node(inner, pos)?;
@@ -224,7 +224,7 @@ impl<'a> Formatter<'a> {
                     self.format_sequence_item(inner, pos)?;
                 }
                 GreenChild::Token { kind, len } => {
-                    self.handle_token(*kind, &self.source[pos..pos + *len]);
+                    self.handle_token(*kind, &self.source[pos..pos + *len as usize]);
                 }
                 GreenChild::Node(inner) => {
                     self.format_node(inner, pos)?;
@@ -242,7 +242,7 @@ impl<'a> Formatter<'a> {
         for child in node.children() {
             match child {
                 GreenChild::Token { kind, len } => {
-                    let text = &self.source[pos..pos + *len];
+                    let text = &self.source[pos..pos + *len as usize];
                     if saw_colon
                         && !matches!(
                             kind,
@@ -291,7 +291,7 @@ impl<'a> Formatter<'a> {
         for child in node.children() {
             match child {
                 GreenChild::Token { kind, len } => {
-                    let text = &self.source[pos..pos + *len];
+                    let text = &self.source[pos..pos + *len as usize];
                     if saw_dash
                         && !matches!(
                             kind,
